+++ src/clip.rs
@@ -27,27 +27,13 @@ pub fn clip(
         clip_canvas.apply_transform(usvg::Transform::from_bbox(bbox).to_native());
     }
 
-    let ts = clip_canvas.transform;
-    for node in cp.root.children() {
-        clip_canvas.apply_transform(node.transform().to_native());
-
-        match *node.borrow() {
-            usvg::NodeKind::Path(ref path_node) => {
-                crate::path::draw(
-                    tree,
-                    path_node,
-                    tiny_skia::BlendMode::Clear,
-                    &mut clip_canvas,
-                );
-            }
-            usvg::NodeKind::Group(ref g) => {
-                clip_group(tree, &node, g, bbox, &mut clip_canvas);
-            }
-            _ => {}
-        }
-
-        clip_canvas.transform = ts;
-    }
+    draw_children(
+        tree,
+        &cp.root,
+        bbox,
+        tiny_skia::BlendMode::Clear,
+        &mut clip_canvas,
+    );
 
     if let Some(ref cp) = cp.clip_path {
         clip(tree, cp, bbox, canvas);
@@ -67,47 +53,68 @@ pub fn clip(
     Some(())
 }
 
+fn draw_children(
+    tree: &usvg::Tree,
+    node: &usvg::Node,
+    bbox: usvg::PathBbox,
+    mode: tiny_skia::BlendMode,
+    canvas: &mut Canvas,
+) {
+    let ts = canvas.transform;
+    for child in node.children() {
+        canvas.apply_transform(child.transform().to_native());
+
+        match *child.borrow() {
+            usvg::NodeKind::Path(ref path_node) => {
+                crate::path::draw(tree, path_node, mode, canvas);
+            }
+            usvg::NodeKind::Group(ref g) => {
+                if let Some(ref cp) = g.clip_path {
+                    // If a `clipPath` child also has a `clip-path`
+                    // then we should render this child on a new canvas,
+                    // clip it, and only then draw it to the `clipPath`.
+                    clip_group(tree, &child, cp, bbox, canvas);
+                } else {
+                    draw_children(tree, &child, bbox, mode, canvas);
+                }
+            }
+            _ => {}
+        }
+
+        canvas.transform = ts;
+    }
+}
+
 fn clip_group(
     tree: &usvg::Tree,
     node: &usvg::Node,
-    g: &usvg::Group,
+    cp: &usvg::ClipPath,
     bbox: usvg::PathBbox,
     canvas: &mut Canvas,
 ) -> Option<()> {
-    if let Some(ref cp) = g.clip_path {
-        // If a `clipPath` child also has a `clip-path`
-        // then we should render this child on a new canvas,
-        // clip it, and only then draw it to the `clipPath`.
+    let mut clip_pixmap = tiny_skia::Pixmap::new(canvas.pixmap.width(), canvas.pixmap.height())?;
+    let mut clip_canvas = Canvas::from(clip_pixmap.as_mut());
+    clip_canvas.transform = canvas.transform;
 
-        let mut clip_pixmap =
-            tiny_skia::Pixmap::new(canvas.pixmap.width(), canvas.pixmap.height())?;
-        let mut clip_canvas = Canvas::from(clip_pixmap.as_mut());
-        clip_canvas.transform = canvas.transform;
+    draw_children(
+        tree,
+        node,
+        bbox,
+        tiny_skia::BlendMode::SourceOver,
+        &mut clip_canvas,
+    );
+    clip(tree, cp, bbox, &mut clip_canvas);
 
-        draw_group_child(tree, node, &mut clip_canvas);
-        clip(tree, cp, bbox, &mut clip_canvas);
-
-        let mut paint = tiny_skia::PixmapPaint::default();
-        paint.blend_mode = tiny_skia::BlendMode::Xor;
-        canvas.pixmap.draw_pixmap(
-            0,
-            0,
-            clip_pixmap.as_ref(),
-            &paint,
-            tiny_skia::Transform::identity(),
-            None,
-        );
-    }
+    let mut paint = tiny_skia::PixmapPaint::default();
+    paint.blend_mode = tiny_skia::BlendMode::Xor;
+    canvas.pixmap.draw_pixmap(
+        0,
+        0,
+        clip_pixmap.as_ref(),
+        &paint,
+        tiny_skia::Transform::identity(),
+        None,
+    );
 
     Some(())
-}
-
-fn draw_group_child(tree: &usvg::Tree, node: &usvg::Node, canvas: &mut Canvas) {
-    if let Some(child) = node.first_child() {
-        canvas.apply_transform(child.transform().to_native());
-
-        if let usvg::NodeKind::Path(ref path_node) = *child.borrow() {
-            crate::path::draw(tree, path_node, tiny_skia::BlendMode::SourceOver, canvas);
-        }
-    }
 }